# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: WardFlow API
  version: "1.0.0"
  description: |
    WardFlow inpatient/ED care-coordination API.

    All endpoints (except /health and /readyz) require a Bearer JWT token.
    RBAC is enforced at the API layer — never rely on frontend-only checks.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.wardflow.internal
    description: Production (internal)
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# TAGS — one per domain module
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      operationId: register
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      summary: Authenticate and receive a JWT
      operationId: login
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      summary: Get current authenticated user
      operationId: getMe
      tags: [Auth]
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Encounters ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable parameters ────────────────────────────────────────────────────
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - field: patientId
                  issue: required field missing
              correlationId: "abc-123"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: Resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    FieldError:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        correlationId:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"

    # Pagination
...
    RegisterRequest:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        unitIds:
          type: array
          items:
            type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserInfo"
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        unitIds:
          type: array
          items:
            type: string
        departmentIds:
          type: array
          items:
            type: string
    UserRole:
      type: string
      enum:
        - nurse
        - provider
        - charge_nurse
        - operations
        - consult
        - transport
        - quality_safety
        - admin

    # Encounter
...